requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2.16
django-cors-headers==4.5.0
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
idna==3.10
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
gunicorn==20.1.0

